small updates

+++ Field.h
@@ -11,7 +11,7 @@ public:
 	void Propagator(); //wave propagation
 	
 private:
-	void initialize(int size, double *A, std::vector<int>  x_range, std::vector<int> val);
+	void initialize(int size, double *A, std::vector<int>  x_range, std::vector<double> val);
 	void print_to_file(int size, double *A, double *pos, int itteration);
